Make DocumentTable class-based
